Refactor: (Feedback) SessionEnrollment 내부에서 원시값을 가지고 도메인 객체 (Capacity, Fee) 를 생성하도록 수정
