+++ src/main/java/nextstep/courses/domain/enrollment/CostSessionEnrollment.java
@@ -1,11 +1,11 @@
 package nextstep.courses.domain.enrollment;
 
-import nextstep.courses.domain.engine.ConcreteSessionEnrollment;
+import nextstep.courses.domain.enrollment.engine.ConcreteSessionEnrollment;
 import nextstep.payments.domain.Payment;
 
 public class CostSessionEnrollment extends ConcreteSessionEnrollment {
 
-    public CostSessionEnrollment(Long id, Long sessionId, SessionStatus status, SessionCapacity capacity, SessionFee fee) {
+    public CostSessionEnrollment(Long id, Long sessionId, SessionStatus status, int capacity, long fee) {
         super(id, sessionId, status, capacity, fee);
     }
