+++ src/main/java/nextstep/courses/domain/enrollment/FreeSessionEnrollment.java
@@ -1,12 +1,12 @@
 package nextstep.courses.domain.enrollment;
 
-import nextstep.courses.domain.engine.ConcreteSessionEnrollment;
+import nextstep.courses.domain.enrollment.engine.ConcreteSessionEnrollment;
 import nextstep.payments.domain.Payment;
 
 public class FreeSessionEnrollment extends ConcreteSessionEnrollment {
 
     public FreeSessionEnrollment(Long id, Long sessionId, SessionStatus status) {
-        super(id, sessionId, status, new SessionCapacity(id, sessionId, SessionCapacity.INFINITY), new SessionFee(id, sessionId, SessionFee.FREE));
+        super(id, sessionId, status, SessionCapacity.INFINITY, SessionFee.FREE);
     }
 
     @Override
