+++ src/test/java/nextstep/courses/domain/SessionEnrollmentTest.java
@@ -1,19 +1,15 @@
 package nextstep.courses.domain;
 
-import nextstep.courses.domain.engine.SessionEnrollment;
-import nextstep.courses.domain.enrollment.SessionCapacity;
-import nextstep.courses.domain.enrollment.SessionFee;
 import nextstep.courses.domain.enrollment.SessionStatus;
 import nextstep.courses.domain.enrollment.Student;
+import nextstep.courses.domain.enrollment.engine.SessionEnrollment;
 import nextstep.payments.domain.Payment;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 import static nextstep.courses.domain.fixture.PaymentFixture.payment;
-import static nextstep.courses.domain.fixture.SessionCapacityFixture.sessionCapacity;
 import static nextstep.courses.domain.fixture.SessionEnrollmentFixture.costSessionEnrollment;
 import static nextstep.courses.domain.fixture.SessionEnrollmentFixture.freeSessionEnrollment;
-import static nextstep.courses.domain.fixture.SessionFeeFixture.sessionFee;
 import static nextstep.courses.domain.fixture.StudentFixture.student;
 import static org.assertj.core.api.Assertions.assertThatNoException;
 
@@ -35,9 +31,7 @@ public class SessionEnrollmentTest {
     @DisplayName("[성공] 유료 강의를 수강신청 한다.")
     void 유료_강의_수강신청() {
         SessionStatus status = SessionStatus.RECRUITING;
-        SessionCapacity capacity = sessionCapacity(10);
-        SessionFee sessionFee = sessionFee(800_000L);
-        SessionEnrollment enrollment = costSessionEnrollment(status, capacity, sessionFee);
+        SessionEnrollment enrollment = costSessionEnrollment(status, 10, 800_000L);
 
         Student student = student();
         Payment payment = payment(800_000L);
